renamed uuids to randomValues
