--- tests/unit/uuids.js
@@ -1,1 +1,0 @@
-export const uuids = ['6f2e0800-a5a4-11eb-bcbc-0242ac130002', '6f2e0a30-a5a4-11eb-bcbc-0242ac130002', '6f2e0b16-a5a4-11eb-bcbc-0242ac130002', '6f2e0cf6-a5a4-11eb-bcbc-0242ac130002', '6f2e0dbe-a5a4-11eb-bcbc-0242ac130002', '6f2e0e7c-a5a4-11eb-bcbc-0242ac130002', '6f2e0fb2-a5a4-11eb-bcbc-0242ac130002', '6f2e1066-a5a4-11eb-bcbc-0242ac130002']
